pyproject.toml
[project]
name = "bnr-converter"
version = "0.1.4"
description = "MCP server for converting foreign currency amounts to RON (Romanian New Leu) using the rates published by BNR (Banca Nationala a Romaniei)"
authors = [
    { name = "Mihai Ionita", email = "mionita1980@proton.me" }
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
 "fastmcp>=2.14.5",
 "starlette>=0.52.1",
 "uvicorn>=0.40.0",
 "requests>=2.32.5",
]

[project.urls]
Homepage = "https://github.com/mionita1980/bnr-converter-mcp-server"
Source = "https://github.com/mionita1980/bnr-converter-mcp-server"

[project.scripts]
bnr-converter = "bnrconverter.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bnrconverter"]
